[Util] Extract access checking functionality
