+++ services/backend/src/services/userService.ts
@@ -6,8 +6,8 @@ import { serviceProvider } from '../config'
 import { roles } from '../config/roles'
 import { sequelizeUser } from '../database/connection'
 import { User } from '../models/user'
-import { Role } from '../shared/types'
-import { DetailedProgrammeRights, ExpandedUser, FormattedUser, IamAccess } from '../types'
+import { DetailedProgrammeRights, Role } from '../shared/types'
+import { ExpandedUser, FormattedUser, IamAccess } from '../types'
 
 import { createLocaleComparator, getFullStudyProgrammeRights, hasFullAccessToStudentData } from '../util'
 import * as jami from '../util/jami'
