+++ services/backend/src/types/index.ts
@@ -1,7 +1,6 @@
 export { CourseWithSubsId, ParsedCourse } from './course'
 export { CreditTypeCode } from './creditTypeCode'
 export { Criteria } from './criteria'
-export { DetailedProgrammeRights } from './detailedProgrammeRights'
 export { DegreeProgrammeType } from './degreeProgrammeType'
 export { EnrollmentState } from './enrollmentState'
 export { EnrollmentType } from './enrollmentType'
