+++ services/backend/src/util/index.ts
@@ -1,6 +1,5 @@
 import { ProgrammeModule } from '../models'
-import { Role } from '../shared/types'
-import { DetailedProgrammeRights } from '../types'
+import { DetailedProgrammeRights, Role } from '../shared/types'
 
 const isObjectWithKey = (obj: unknown, key: string): obj is Record<string, unknown> => {
   return typeof obj === 'object' && obj !== null && key in obj
