+++ services/frontend/src/components/PopulationDetails/index.jsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react'
 
-import { getFullStudyProgrammeRights } from '@/common'
 import { populationStatisticsToolTips } from '@/common/InfoToolTips'
 import { PanelView } from '@/components/common/PanelView'
 import { CreditAccumulationGraphHighCharts } from '@/components/CreditAccumulationGraphHighCharts'
@@ -10,6 +9,7 @@ import { InfoBox } from '@/components/InfoBox'
 import { PopulationStudentsContainer as PopulationStudents } from '@/components/PopulationStudents'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
 import { useGetProgressCriteriaQuery } from '@/redux/programmeProgressCriteria'
+import { getFullStudyProgrammeRights } from '@/util/access'
 import { AgeStats } from './AgeStats'
 import { CreditGainStats } from './CreditGainStats'
 import { CourseTableModeSelector } from './CurriculumPicker'
