+++ services/frontend/src/components/Routes/ProtectedRoute.tsx
@@ -1,8 +1,9 @@
 import { Outlet, useLocation } from 'react-router'
-import { checkUserAccess } from '@/common'
+
 import { hasFullAccessToTeacherData } from '@/components/Teachers/util'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
 import { Role } from '@/shared/types'
+import { checkUserAccess } from '@/util/access'
 import { AccessDeniedMessage } from './AccessDeniedMessage'
 
 interface ProtectedRouteProps {
