+++ services/frontend/src/components/StudyProgramme/index.jsx
@@ -2,12 +2,13 @@ import { useState, useCallback } from 'react'
 import { useParams, useLocation, useNavigate } from 'react-router'
 import { Container, Header, Menu, Segment, Tab } from 'semantic-ui-react'
 
-import { getFullStudyProgrammeRights, getUnifiedProgrammeName, isDefaultServiceProvider } from '@/common'
+import { getUnifiedProgrammeName, isDefaultServiceProvider } from '@/common'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { useSemanticTabs } from '@/hooks/tabs'
 import { useTitle } from '@/hooks/title'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
 import { useGetProgrammesQuery } from '@/redux/populations'
+import { getFullStudyProgrammeRights } from '@/util/access'
 import { BasicOverview } from './BasicOverview'
 import { DegreeCoursesTable } from './DegreeCourses'
 import { ProgrammeCourses } from './ProgrammeCourses'
