+++ services/frontend/src/components/Teachers/TeacherStatistics/index.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import ReactMarkdown from 'react-markdown'
 import { Button, Dropdown, Form, Message, Segment } from 'semantic-ui-react'
 
-import { createLocaleComparator, getCurrentSemester, getFullStudyProgrammeRights } from '@/common'
+import { createLocaleComparator, getCurrentSemester } from '@/common'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { TeacherStatisticsTable } from '@/components/Teachers/TeacherStatisticsTable'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
@@ -10,6 +10,7 @@ import { useGetProvidersQuery } from '@/redux/providers'
 import { useGetSemestersQuery } from '@/redux/semesters'
 import { useLazyGetTeacherStatisticsQuery } from '@/redux/teachers'
 import { mapToProviders } from '@/shared/util'
+import { getFullStudyProgrammeRights } from '@/util/access'
 import { hasFullAccessToTeacherData } from '../util'
 
 export const TeacherStatistics = () => {
