+++ services/frontend/src/components/material/NavigationBar/index.tsx
@@ -2,9 +2,10 @@ import { AppBar, Box, Container, Tab, Tabs, Toolbar } from '@mui/material'
 import { useState, useEffect } from 'react'
 import { Link, useLocation } from 'react-router'
 
-import { checkUserAccess, getFullStudyProgrammeRights, isDefaultServiceProvider } from '@/common'
+import { isDefaultServiceProvider } from '@/common'
 import { hasFullAccessToTeacherData } from '@/components/Teachers/util'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
+import { checkUserAccess, getFullStudyProgrammeRights } from '@/util/access'
 import { NavigationButton } from './NavigationButton'
 import { NavigationItem, navigationItems } from './navigationItems'
 import { OodikoneLogo } from './OodikoneLogo'
