+++ services/frontend/src/pages/CourseStatistics/index.tsx
@@ -4,7 +4,6 @@ import { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useLocation } from 'react-router'
 
-import { checkUserAccess, getFullStudyProgrammeRights } from '@/common'
 import { PageTitle } from '@/components/material/PageTitle'
 import { AccessDeniedMessage } from '@/components/Routes/AccessDeniedMessage'
 import { useProgress } from '@/hooks/progress'
@@ -13,7 +12,7 @@ import { useTitle } from '@/hooks/title'
 import { RootState } from '@/redux'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
 import { getCourseStats } from '@/redux/courseStats'
-import { userHasAccessToAllCourseStats } from './courseStatisticsUtils'
+import { checkUserAccess, getFullStudyProgrammeRights, userHasAccessToAllCourseStats } from '@/util/access'
 import { FacultyStatisticsTab } from './FacultyStatisticsTab'
 import { SearchForm } from './SearchForm'
 import { SingleCourseTab } from './SingleCourseTab'
