+++ services/shared/types/detailedProgrammeRights.ts
@@ -1,5 +1,5 @@
 export type DetailedProgrammeRights = {
   code: string
-  limited: boolean
   isIamBased: boolean
+  limited: boolean
 }
